Update nonacute codes
NY has asked that Acute IP be considered anything no SNF (i31).  Since we limit the data to PRM_Line = i and <> i31 this extra macro is not really needed to determine acute/nonacute; however, we may be asked to change in in the future so we will keep the macro for now.