Only exclude beneficiary_status = 'Unknown'
Modified the code to only exclude members with 0 member months from the
avg_age and risk_scr_avg calculation if they are in the unknown
category.  Do this because there are many members with 0 member months
in other categories.  I am unsure what impact excluding everyone with 0
member months would have on the metrics and don't want to affect them.