Added limiting pattern and logic due to CMS change
Do this because CMS has introduced the word "Prospective" into QASSGN Headers which was a keyword check for HASSGN files.  The new logic introduces a new handler after the check for prospective.  If the word "Quarter" is found in the same row that "Prospective" is identified, then the prospective_flag is reverted to False