Update recon step
Do this to ensure both community and new enrollee scores reconcile for every member. Additionally, add handling for an edge case in the new enrollee scores that causes them to be equal to 0.01 with no coefficients if the member's birthdate occurred after the start of the risk score training period.